confirm HUI workflow works
